[refactor] move NarwhalManager to its own file.
