+++ crates/sui-core/src/unit_tests/narwhal_manager_tests.rs
@@ -2,7 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::authority::test_authority_builder::TestAuthorityBuilder;
-use crate::consensus_manager::{NarwhalConfiguration, NarwhalManager, NarwhalManagerMetrics};
+use crate::consensus_manager::narwhal_manager::{
+    NarwhalConfiguration, NarwhalManager, NarwhalManagerMetrics,
+};
 use bytes::Bytes;
 use fastcrypto::bls12381;
 use fastcrypto::traits::KeyPair;
